fixed warning for 64bit architectures.

+++ Tappy Plane/TPPlane.m
@@ -126,7 +126,7 @@ static const CGFloat kTPMaxAltitude = 300.0;
 -(void)setRandomColor
 {
     [self removeActionForKey:kTPKeyPlaneAnimation];
-    SKAction *animation = [self.planeAnimations objectAtIndex:arc4random_uniform(self.planeAnimations.count)];
+    SKAction *animation = [self.planeAnimations objectAtIndex:arc4random_uniform((uint)self.planeAnimations.count)];
     [self runAction:animation withKey:kTPKeyPlaneAnimation];
     
     if (!self.engineRunning) {
